% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textreadr_class.R
\name{print.textreadr}
\alias{print.textreadr}
\title{Prints a textreadr Object}
\usage{
\method{print}{textreadr}(x, ...)
}
\arguments{
\item{x}{A \code{\link[base]{data.frame}} textreadr object.}

\item{\ldots}{ignored.}
}
\description{
Prints a textreadr object
}

