% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textreadr-package.R
\docType{package}
\name{textreadr}
\alias{package-textreadr}
\alias{textreadr}
\alias{textreadr-package}
\title{What the Package Does (one line)}
\description{
What the package does (one or more paragraphs)
}

