% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download Documents}
\usage{
download(url, loc = tempdir())
}
\arguments{
\item{url}{The download url(s).}

\item{loc}{Where to put the files.}
}
\value{
Places a copy of the downloaded document in location specified and returns
vector of the locations as string paths.
}
\description{
This function enables downloading documents.
}
\examples{
\dontrun{
m <- download(
c('https://cran.r-project.org/web/packages/curl/curl.pdf',
"https://github.com/trinker/textreadr/raw/master/inst/docs/rl10075oralhistoryst002.pdf")
)

m
}
}

