% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textreadr_class.R
\name{print.textreadr}
\alias{print.textreadr}
\title{Prints a textreadr Object}
\usage{
\method{print}{textreadr}(x, width = 40, ...)
}
\arguments{
\item{x}{A \code{\link[base]{data.frame}} textreadr object.}

\item{width}{The width of the columns to be displayed.}

\item{\ldots}{Other arguments passed to \code{\link[textreadr]{peek}}.}
}
\description{
Prints a textreadr object
}

