% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_doc.R
\name{read_doc}
\alias{read_doc}
\title{Read in .doc Content}
\usage{
read_doc(file, skip = 0, antiword.path = "C:/antiword/antiword.exe")
}
\arguments{
\item{file}{The path to the .doc file.}

\item{skip}{The number of lines to skip.}

\item{antiword.path}{path to \href{http://www.winfield.demon.nl}{antiword}}
}
\value{
Returns a character vector.
}
\description{
Read in the content from a .doc file using \href{http://www.winfield.demon.nl}{antiword}.
}
\examples{
\dontrun{
x <- "https://syllabus.chaminade.edu/get_syllabus.php?id=20715"
out <- curl::curl_download(x, tempfile())
(txt <- read_doc(out))
}
}
\keyword{doc}

