% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{read_docx}
\alias{read_docx}
\title{Read in .docx Content}
\usage{
read_docx(file, skip = 0)
}
\arguments{
\item{file}{The path to the .docx file.}

\item{skip}{The number of lines to skip.}
}
\value{
Returns a character vector.
}
\description{
Read in the content from a .docx file.
}
\examples{
\dontrun{
download("http://umlreading.weebly.com/uploads/2/5/2/5/25253346/whole_language_timeline-updated.docx")
(txt <- read_docx("whole_language_timeline-updated.docx"))
}
}
\author{
Bryan Goodrich and Tyler Rinker <tyler.rinker@gmail.com>.
}
\keyword{docx}

