% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docx.R
\name{read_docx}
\alias{read_docx}
\title{Read in .docx Content}
\usage{
read_docx(file, skip = 0)
}
\arguments{
\item{file}{The path to the .docx file.}

\item{skip}{The number of lines to skip.}
}
\value{
Returns a character vector.
}
\description{
Read in the content from a .docx file.
}
\examples{
\dontrun{
x <- "http://www.cybersmart.gov.au/~/media/9999BCDEA99F40DD8170AAD978C8D2F9.docx"
download(x)
(txt <- read_docx("whole_language_timeline-updated.docx"))
}
}
\author{
Bryan Goodrich and Tyler Rinker <tyler.rinker@gmail.com>.
}
\keyword{docx}

