% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_transcript.R
\name{read_transcript}
\alias{read_transcript}
\title{read_transcripts Into R}
\usage{
read_transcript(file, col.names = NULL, text.var = NULL,
  merge.broke.tot = TRUE, header = FALSE, dash = "", ellipsis = "...",
  quote2bracket = FALSE, rm.empty.rows = TRUE, na.strings = c("999", "NA",
  "", " "), sep = NULL, skip = 0, text, comment.char = "", ...)
}
\arguments{
\item{file}{The name of the file which the data are to be read from. Each row
of the table appears as one line of the file. If it does not contain an
absolute path, the file name is relative to the current working directory,
\code{getwd()}.}

\item{col.names}{A character vector specifying the column names of the
transcript columns.}

\item{text.var}{A character string specifying the name of the text variable
will ensure that variable is classed as character.  If \code{NULL}
\code{\link[textreadr]{read_transcript}} attempts to guess the text.variable
(dialogue).}

\item{merge.broke.tot}{logical.  If \code{TRUE} and if the file being read in
is .docx with broken space between a single turn of talk read_transcript
will attempt to merge these into a single turn of talk.}

\item{header}{logical.  If \code{TRUE} the file contains the names of the
variables as its first line.}

\item{dash}{A character string to replace the en and em dashes special
characters (default is to remove).}

\item{ellipsis}{A character string to replace the ellipsis special characters
(default is text ...).}

\item{quote2bracket}{logical. If \code{TRUE} replaces curly quotes with curly
braces (default is \code{FALSE}).  If \code{FALSE} curly quotes are removed.}

\item{rm.empty.rows}{logical.  If \code{TRUE}
\code{\link[textreadr]{read_transcript}}  attempts to remove empty rows.}

\item{na.strings}{A vector of character strings which are to be interpreted
as \code{NA} values.}

\item{sep}{The field separator character. Values on each line of the file are
separated by this character.  The default of \code{NULL} instructs
\code{\link[textreadr]{read_transcript}} to use a separator suitable for the file
type being read in.}

\item{skip}{Integer; the number of lines of the data file to skip before
beginning to read data.}

\item{text}{Character string: if file is not supplied and this is, then data
are read from the value of text. Notice that a literal string can be used to
include (small) data sets within R code.}

\item{comment.char}{A character vector of length one containing a single
character or an empty string. Use \code{""} to turn off the interpretation of
comments altogether.}

\item{\ldots}{Further arguments to be passed to \code{\link[utils]{read.table}}.}
}
\value{
Returns a dataframe of dialogue and people.
}
\description{
Read .docx, .csv, .xlsx, .xlsx, or .txt transcript style files into R.
}
\note{
If a transcript is a .docx file read_transcript expects two columns
(generally person and dialogue) with some sort of separator (default is colon
separator).  .doc files must be converted to .docx before reading in.
}
\section{Warning}{
 \code{\link[textreadr]{read_transcript}} may contain errors if the
file being read in is .docx.  The researcher should carefully investigate
each transcript for errors before further parsing the data.
}
\author{
Bryan Goodrich and Tyler Rinker <tyler.rinker@gmail.com>.
}
\references{
\url{https://github.com/trinker/qdap/wiki/Reading-.docx-\%5BMS-Word\%5D-Transcripts-into-R}
}
\keyword{transcript}

