% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textreadr-package.R
\docType{package}
\name{textreadr}
\alias{package-textreadr}
\alias{textreadr}
\alias{textreadr-package}
\title{Read Text Documents into R}
\description{
A small collection of convenience tools for reading text documents into R.
}

